create new package todmap for pysm operator (#204)
* create new package scan for pysm operator

closes #200

* rename scan to todmap

* add todmap to build system

* move pysm operator functions to todmap

* fix import in tests
