+++ configure.ac
@@ -232,6 +232,7 @@ AC_CONFIG_FILES([src/libtoast/test/Makefile])
 AC_CONFIG_FILES([src/libtoast/capi/Makefile])
 AC_CONFIG_FILES([src/python/Makefile])
 AC_CONFIG_FILES([src/python/tod/Makefile])
+AC_CONFIG_FILES([src/python/todmap/Makefile])
 AC_CONFIG_FILES([src/python/fod/Makefile])
 AC_CONFIG_FILES([src/python/map/Makefile])
 AC_CONFIG_FILES([src/python/tests/Makefile])
