+++ src/python/todmap/Makefile.am
@@ -1,0 +1,28 @@
+SUBDIRS = 
+
+# Initialize compile and linking flags
+
+AM_CPPFLAGS = \
+-I$(top_srcdir)/src/libtoast \
+$(PYTHON_CPPFLAGS)
+
+AM_CFLAGS =
+AM_CXXFLAGS =
+
+AM_LDFLAGS = $(PYTHON_LDFLAGS)
+
+AM_LIBS = $(PYTHON_LIBS)
+
+# Append compile and linking flags based on configure variables
+include $(top_srcdir)/src/build_options.am
+
+
+toasttodmapdir = $(pythondir)/toast/todmap
+
+toasttodmap_PYTHON = \
+__init__.py \
+pysm_operator.py
+
+clean-local :
+	@rm -f *.pyc
+
